fixed a bunch of minor server bugs

+++ src/main/java/de/ellpeck/rockbottom/net/packet/toclient/PacketEntityUpdate.java
@@ -12,10 +12,10 @@ import java.util.UUID;
 public class PacketEntityUpdate implements IPacket{
 
     private UUID uniqueId;
-    private double x;
-    private double y;
-    private double motionX;
-    private double motionY;
+    private float x;
+    private float y;
+    private float motionX;
+    private float motionY;
     private Direction facing;
     private boolean collidedHor;
     private boolean collidedVert;
@@ -23,10 +23,10 @@ public class PacketEntityUpdate implements IPacket{
 
     public PacketEntityUpdate(UUID uniqueId, double x, double y, double motionX, double motionY, Direction facing, boolean collidedHor, boolean collidedVert, boolean onGround){
         this.uniqueId = uniqueId;
-        this.x = x;
-        this.y = y;
-        this.motionX = motionX;
-        this.motionY = motionY;
+        this.x = (float)x;
+        this.y = (float)y;
+        this.motionX = (float)motionX;
+        this.motionY = (float)motionY;
         this.facing = facing;
         this.collidedHor = collidedHor;
         this.collidedVert = collidedVert;
@@ -41,10 +41,10 @@ public class PacketEntityUpdate implements IPacket{
     public void toBuffer(ByteBuf buf){
         buf.writeLong(this.uniqueId.getMostSignificantBits());
         buf.writeLong(this.uniqueId.getLeastSignificantBits());
-        buf.writeDouble(this.x);
-        buf.writeDouble(this.y);
-        buf.writeDouble(this.motionX);
-        buf.writeDouble(this.motionY);
+        buf.writeFloat(this.x);
+        buf.writeFloat(this.y);
+        buf.writeFloat(this.motionX);
+        buf.writeFloat(this.motionY);
         buf.writeInt(this.facing.ordinal());
         buf.writeBoolean(this.collidedHor);
         buf.writeBoolean(this.collidedVert);
@@ -54,10 +54,10 @@ public class PacketEntityUpdate implements IPacket{
     @Override
     public void fromBuffer(ByteBuf buf){
         this.uniqueId = new UUID(buf.readLong(), buf.readLong());
-        this.x = buf.readDouble();
-        this.y = buf.readDouble();
-        this.motionX = buf.readDouble();
-        this.motionY = buf.readDouble();
+        this.x = buf.readFloat();
+        this.y = buf.readFloat();
+        this.motionX = buf.readFloat();
+        this.motionY = buf.readFloat();
         this.facing = Direction.DIRECTIONS[buf.readInt()];
         this.collidedHor = buf.readBoolean();
         this.collidedVert = buf.readBoolean();
